fix: use Int64 flag type for int64 flags instead of Int
Fixes pre-existing bug where gh-app-id and gh-app-installation-id were
registered as Int flags instead of Int64 flags, which could cause issues
with large ID values.

+++ cmd/server.go
@@ -823,7 +823,7 @@ func (s *ServerCmd) Init() *cobra.Command {
 			usage = fmt.Sprintf("%s (default %d)", usage, f.defaultValue)
 			s.Viper.SetDefault(name, f.defaultValue)
 		}
-		c.Flags().Int(name, 0, usage+"\n")
+		c.Flags().Int64(name, 0, usage+"\n")
 		if f.hidden {
 			c.Flags().MarkHidden(name) // nolint: errcheck
 		}
